+++ LibCarla/source/carla/client/detail/Client.h
@@ -88,7 +88,6 @@ namespace detail {
 
     void DestroyTrafficManager(uint16_t port) const;
 
-
     void SetTimeout(time_duration timeout);
 
     time_duration GetTimeout() const;
@@ -241,6 +240,24 @@ namespace detail {
     geom::Transform GetActorComponentRelativeTransform(
         rpc::ActorId actor,
         const std::string componentName);
+
+    std::vector<geom::Transform> GetActorBoneWorldTransforms(
+        rpc::ActorId actor);
+    
+    std::vector<geom::Transform> GetActorBoneRelativeTransforms(
+        rpc::ActorId actor);
+
+    std::vector<std::string> GetActorComponentNames(
+        rpc::ActorId actor);         
+
+    std::vector<std::string> GetActorBoneNames(
+        rpc::ActorId actor);
+
+    std::vector<geom::Transform> GetActorSocketWorldTransforms(
+        rpc::ActorId actor);
+
+    std::vector<geom::Transform> GetActorSocketRelativeTransforms(
+        rpc::ActorId actor);
 
     void SetActorSimulatePhysics(
         rpc::ActorId actor,